Refactored some of the common parts for new_environ.
