+++ src/options/new_environ/client.cpp
@@ -1,4 +1,6 @@
 #include "telnetpp/options/new_environ/client.hpp"
+#include "telnetpp/options/new_environ/detail/request_parser.hpp"
+#include "telnetpp/options/new_environ/detail/stream.hpp"
 #include "telnetpp/options/new_environ.hpp"
 
 namespace telnetpp { namespace options { namespace new_environ {
@@ -6,90 +8,12 @@ namespace telnetpp { namespace options { namespace new_environ {
 namespace {
 
 // ==========================================================================
-// PARSE_REQUESTS
-// ==========================================================================
-std::vector<request> parse_requests(u8stream const &data)
-{
-    std::vector<request> requests;
-    request current_request;
-    
-    enum class state
-    {
-        send,
-        var_or_uservar,
-        name,
-        name_esc
-    };
-    
-    state current_state = state::send;
-    
-    for (auto &byte : data)
-    {
-        switch (current_state)
-        {
-            case state::send :
-                if (byte != send)
-                {
-                    return {};
-                }
-                
-                current_state = state::var_or_uservar;
-                break;
-                
-            case state::var_or_uservar :
-                if (byte == var || byte == uservar)
-                {
-                    current_request.type = byte;
-                    current_state = state::name;
-                }
-                else
-                {
-                    return {};
-                }
-                break;
-                
-            case state::name :
-                if (byte == var || byte == uservar)
-                {
-                    requests.push_back(current_request);
-                    current_request.name  = "";
-                    current_request.type  = byte;
-                }
-                else if (byte == esc)
-                {
-                    current_state = state::name_esc;
-                }
-                else
-                {
-                    current_request.name.push_back(byte);
-                }
-                break;
-                
-            case state::name_esc :
-                current_request.name.push_back(byte);
-                current_state = state::name;
-                break;
-                
-            default :
-                return {};
-        }
-    }
-    
-    if (!current_request.name.empty())
-    {
-        requests.push_back(current_request);
-    }
-    
-    return requests;
-}
-
-// ==========================================================================
 // VARIABLE_WAS_REQUESTED
 // ==========================================================================
 bool variable_was_requested(
-    std::vector<request>                      const &requests,
-    telnetpp::u8                                     variable_type,
-    std::pair<std::string const, std::string> const &variable)
+    std::vector<request> const &requests,
+    telnetpp::u8                variable_type,
+    std::string          const &name)
 {
     if (requests.empty())
     {
@@ -99,39 +23,13 @@ bool variable_was_requested(
     for (auto const &request : requests)
     {
         if (request.type == variable_type
-         && request.name == variable.first)
+         && request.name == name)
         {
             return true;
         }
     }
     
     return false;
-}
-
-// ==========================================================================
-// APPEND_STRING
-// ==========================================================================
-u8stream &append_string(
-    u8stream          &stream,
-    std::string const &text)
-{
-    for (auto ch : text)
-    {
-        switch (ch)
-        {
-            case var     : // Fall-through
-            case value   : // Fall-through
-            case esc     : // Fall-through
-            case uservar : // Fall-through
-                stream.push_back(esc);
-                // Fall-through
-            default :
-                stream.push_back(ch);
-                break;
-        }
-    }
-
-    return stream;
 }
 
 // ==========================================================================
@@ -146,9 +44,9 @@ u8stream &append_variable(
     std::string const &val  = variable.second;
     
     stream.push_back(type);
-    append_string(stream, name);
+    detail::append_escaped(stream, name);
     stream.push_back(value);
-    append_string(stream, val);
+    detail::append_escaped(stream, val);
     
     return stream;    
 }
@@ -187,7 +85,7 @@ std::vector<telnetpp::token> client::delete_variable(std::string const &name)
     variables_.erase(name);
     
     u8stream result = { info, var };
-    append_string(result, name);
+    detail::append_escaped(result, name);
     
     return { telnetpp::element{
         telnetpp::subnegotiation(option(), result)
@@ -219,7 +117,7 @@ std::vector<telnetpp::token> client::delete_user_variable(
     user_variables_.erase(name);
     
     u8stream result = { info, uservar };
-    append_string(result, name);
+    detail::append_escaped(result, name);
     
     return { telnetpp::element{
         telnetpp::subnegotiation(option(), result)
@@ -232,13 +130,13 @@ std::vector<telnetpp::token> client::delete_user_variable(
 std::vector<telnetpp::token> client::handle_subnegotiation(
     u8stream const &content)
 {
-    auto const &requests = parse_requests(content);
+    auto const &requests = detail::parse_requests(content);
     
     u8stream result = { is };
     
     for (auto &variable : variables_)
     {
-        if (variable_was_requested(requests, var, variable))
+        if (variable_was_requested(requests, var, variable.first))
         {
             append_variable(result, var, variable);
         }
@@ -246,7 +144,7 @@ std::vector<telnetpp::token> client::handle_subnegotiation(
 
     for (auto &variable : user_variables_)
     {
-        if (variable_was_requested(requests, uservar, variable))
+        if (variable_was_requested(requests, uservar, variable.first))
         {
             append_variable(result, uservar, variable);
         }
